PRSD-NONE: Moves assertIsPage into BasePage companion object
